feat: extract rememberSyntaxHighlightedEditorValue() from SyntaxHighlightedTextEditor
Mirrors the existing rememberHighlightedCode / rememberHighlightedCodeBothThemes
pattern: the debounce + engine pipeline moves into a standalone @Composable helper
in RememberHelpers.kt, leaving SyntaxHighlightedTextEditor as a thin layout shell.

- rememberSyntaxHighlightedEditorValue(value, language, theme, debounceMs,
  onHighlightComplete): State<TextFieldValue>
  - owns HighlightSnapshot state, rememberUpdatedState wrappers,
    LaunchedEffect with debounce, and span-clipping when block
  - callers can use it with any text field (BasicTextField, OutlinedTextField,
    third-party editor) without forking SyntaxHighlightedTextEditor
  - easier to test in isolation: no Surface/BasicTextField needed in the tree

- SyntaxHighlightedTextEditor now calls rememberSyntaxHighlightedEditorValue
  and renders Surface { BasicTextField(...) } (~30 lines of layout code)

- HighlightSnapshot private data class moved from SyntaxHighlightedTextEditor.kt
  to RememberHelpers.kt (private, same file as its only consumer)

No behavior change for existing callers of SyntaxHighlightedTextEditor.
