+++ compose-highlight/src/main/kotlin/dev/hossain/highlight/ui/RememberHelpers.kt
@@ -4,17 +4,21 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.State
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.input.TextFieldValue
 import dev.hossain.highlight.engine.HighlightEngine
 import dev.hossain.highlight.engine.HighlightException
 import dev.hossain.highlight.engine.HighlightResult
 import dev.hossain.highlight.engine.HighlightTheme
 import dev.hossain.highlight.engine.ThemedHighlightResult
+import kotlinx.coroutines.delay
 
 /**
  * Creates and remembers a [HighlightEngine] scoped to the composition.
@@ -250,4 +254,135 @@ fun rememberHighlightedCodeBothThemes(
     }
 
     return state
+}
+
+/**
+ * Holds the result of a syntax-highlight call together with the [language] and [theme] that
+ * produced it. Stored as local state so the composable can detect in-composition whether the
+ * cached result is still valid for the current language/theme, eliminating the need for a
+ * separate `LaunchedEffect` that resets state asynchronously.
+ */
+private data class HighlightSnapshot(
+    val annotated: AnnotatedString,
+    val language: String,
+    val theme: HighlightTheme,
+)
+
+/**
+ * Runs the debounce + syntax-highlight pipeline for a live code editor and returns the
+ * display [TextFieldValue] ready to pass directly to `BasicTextField` (or any other text
+ * field that accepts [TextFieldValue]).
+ *
+ * This is the lower-level counterpart to [SyntaxHighlightedTextEditor]: it owns the engine
+ * call, the debounce window, and the span-clipping logic, but does **not** render anything.
+ * Callers drive their own layout, which makes it easy to:
+ * - wrap an `OutlinedTextField` or a third-party editor with syntax highlighting
+ * - test the highlight pipeline in isolation without a `Surface`/`BasicTextField` in the tree
+ *
+ * The returned [State] is updated each time a new highlight result arrives. Between updates
+ * the previous spans are preserved and clipped to the current text length, so the editor
+ * never flickers or loses color while the user is typing.
+ *
+ * ## Usage - standalone (with BasicTextField)
+ *
+ * ```kotlin
+ * var editorValue by remember { mutableStateOf(TextFieldValue("fun hello() = println(\"Hello!\")")) }
+ *
+ * HighlightThemeProvider(
+ *     lightHighlightTheme = HighlightTheme.tomorrow(),
+ *     darkHighlightTheme  = HighlightTheme.tomorrowNight(),
+ * ) {
+ *     val displayValue by rememberSyntaxHighlightedEditorValue(
+ *         value    = editorValue,
+ *         language = "kotlin",
+ *     )
+ *     BasicTextField(
+ *         value         = displayValue,
+ *         onValueChange = { editorValue = it },
+ *     )
+ * }
+ * ```
+ *
+ * ## Usage - with SyntaxHighlightedTextEditor (preferred for most cases)
+ *
+ * For the common case of a themed, bordered editor use [SyntaxHighlightedTextEditor] directly.
+ * It calls this function internally and adds the `Surface` + padding + test-tag wrapper.
+ *
+ * @param value The current [TextFieldValue], including text, cursor position, and selection.
+ * @param language Highlight.js language identifier (e.g. `"kotlin"`, `"python"`, `"sql"`).
+ * @param theme The highlight theme to apply. Defaults to [LocalHighlightTheme].
+ * @param debounceMs Milliseconds to wait after the last keystroke before triggering a new
+ *   highlight call. Defaults to 150 ms. If this value changes at runtime the new delay is
+ *   picked up on the next highlight cycle without restarting the effect.
+ * @param onHighlightComplete Optional callback invoked each time a highlight cycle completes
+ *   successfully. Receives the resulting [AnnotatedString] with syntax spans applied. Useful
+ *   for testing (wait until the first result arrives) and for observing the highlight output
+ *   without owning the editor's text state.
+ * @return A [State] holding a [TextFieldValue] with syntax-highlight spans applied, preserving
+ *   the original cursor position and selection. Falls back to plain text while a highlight
+ *   result is in flight, on error, or when the language/theme has just changed.
+ */
+@ExperimentalHighlightApi
+@Composable
+fun rememberSyntaxHighlightedEditorValue(
+    value: TextFieldValue,
+    language: String,
+    theme: HighlightTheme = LocalHighlightTheme.current,
+    debounceMs: Long = 150L,
+    onHighlightComplete: ((AnnotatedString) -> Unit)? = null,
+): State<TextFieldValue> {
+    val engine = rememberHighlightEngine()
+    var highlighted by remember { mutableStateOf<HighlightSnapshot?>(null) }
+    // rememberUpdatedState ensures a changed debounceMs or callback is used by the running
+    // effect without restarting it (restarting would reset the debounce window mid-keystroke).
+    val currentDebounceMs by rememberUpdatedState(debounceMs)
+    val currentOnHighlightComplete by rememberUpdatedState(onHighlightComplete)
+
+    // Re-highlight with debounce whenever the text, language, or theme changes.
+    // LaunchedEffect cancels the previous coroutine on each change, so rapid keystrokes
+    // naturally coalesce into a single highlight call after the user pauses.
+    LaunchedEffect(value.text, language, theme) {
+        delay(currentDebounceMs)
+        engine
+            .highlight(value.text, language, theme)
+            .onSuccess { result ->
+                highlighted = HighlightSnapshot(result.annotated, language, theme)
+                currentOnHighlightComplete?.invoke(result.annotated)
+            }.onFailure { highlighted = null }
+    }
+
+    // Merge highlight spans into the TextFieldValue while preserving cursor and selection.
+    //
+    // Three cases:
+    // 1. No snapshot yet, or snapshot is stale (different language/theme) - show plain text.
+    //    Stale detection is in-composition: no separate LaunchedEffect needed to clear state.
+    // 2. Snapshot text exactly matches current text - apply spans directly (steady state).
+    // 3. Text has changed since the last snapshot (user is typing, debounce pending) -
+    //    clip old spans to the new text length and apply them. Spans before any edit point
+    //    remain correctly colored; only the newly typed characters briefly have no span.
+    //    This gives the illusion of live highlighting while the debounce window is open.
+    val currentText = value.text
+    val snapshot = highlighted
+    val annotated =
+        when {
+            snapshot == null || snapshot.language != language || snapshot.theme != theme -> {
+                AnnotatedString(currentText)
+            }
+
+            snapshot.annotated.text == currentText -> {
+                snapshot.annotated
+            }
+
+            else -> {
+                val builder = AnnotatedString.Builder(currentText)
+                snapshot.annotated.spanStyles.forEach { range ->
+                    val start = range.start.coerceAtMost(currentText.length)
+                    val end = range.end.coerceAtMost(currentText.length)
+                    if (start < end) builder.addStyle(range.item, start, end)
+                }
+                builder.toAnnotatedString()
+            }
+        }
+
+    return remember(value, annotated) { mutableStateOf(value.copy(annotatedString = annotated)) }
 }
